[pkg/controlplane/routes]: refactor controlplane routes
1. make routes inside of controlplane
2. rename `master -> controlplane`
