+++ pkg/BUILD
@@ -69,7 +69,6 @@ filegroup(
         "//pkg/proxy:all-srcs",
         "//pkg/quota:all-srcs",
         "//pkg/registry:all-srcs",
-        "//pkg/routes:all-srcs",
         "//pkg/scheduler:all-srcs",
         "//pkg/security:all-srcs",
         "//pkg/securitycontext:all-srcs",
