+++ pkg/controlplane/routes/logs.go
