+++ pkg/controlplane/routes/openidmetadata.go
